+++ test/unit/replication-couchdb.test.ts
@@ -265,7 +265,10 @@ describe('replication-couchdb.test.ts', () => {
             const c2 = await humansCollection.create(0);
 
             const replicationState1 = await syncLive(c1, server);
+            ensureReplicationHasNoErrors(replicationState1);
             const replicationState2 = await syncLive(c2, server);
+            ensureReplicationHasNoErrors(replicationState2);
+
             const awaitInSync = () => Promise.all([
                 replicationState1.awaitInSync(),
                 replicationState2.awaitInSync()
@@ -285,31 +288,30 @@ describe('replication-couchdb.test.ts', () => {
 
             // wait until it is on the server
             await waitUntil(async () => {
-                const serverDocs = await getAllServerDocs(server.url);
-                return serverDocs.length === 1;
+                const serverDocsInner = await getAllServerDocs(server.url);
+                return serverDocsInner.length === 1;
             });
 
             const endResult = await foundPromise;
             assert.strictEqual(endResult[0].passportId, 'foobar');
 
             const doc1 = await c1.findOne().exec(true);
-            const doc2 = await c1.findOne().exec(true);
+            const doc2 = await c2.findOne().exec(true);
 
             // edit on one side
             await doc1.incrementalPatch({ age: 20 });
             await awaitInSync();
-            assert.strictEqual(doc2.getLatest().age, 20);
+            await waitUntil(() => doc2.getLatest().age === 20);
 
             // edit on one side again
             await doc1.incrementalPatch({ age: 21 });
             await awaitInSync();
-            assert.strictEqual(doc2.getLatest().age, 21);
-
+            await waitUntil(() => doc2.getLatest().age === 21);
 
             // edit on other side
             await doc2.incrementalPatch({ age: 22 });
             await awaitInSync();
-            assert.strictEqual(doc1.getLatest().age, 22);
+            await waitUntil(() => doc1.getLatest().age === 22);
 
             c1.database.destroy();
             c2.database.destroy();
@@ -318,8 +320,6 @@ describe('replication-couchdb.test.ts', () => {
     });
     describe('ISSUES', () => {
         it('#4299 CouchDB push is throwing error because of missing revision', async () => {
-
-
             const server = await SpawnServer.spawn();
 
             // create a collection